LA-117 Migrate JavaScript ES5 to ES6 - Rename file extension js into mjs for module files.
