+++ ArticleShapeExtractor/modules/BrandSectionMapResolver.mjs
