+++ ArticleShapeExtractor/modules/PreferencesManager.mjs
