+++ ArticleShapeExtractor/modules/VersionUtils.mjs
